Finish tests refactor
